Changes so that when a new query is submitted, the tables containing the previous result are cleared.  This should prevent a reoccurrence of fault 20030313.004
git-svn-id: file:///jac_sw/svnroot/scicom/trunk/omp/java/qt@5695 64693ad5-8cdc-4012-857f-4cd9bf50b8c8

+++ src/main/edu/jach/qt/gui/ProgramTree.java
@@ -1280,7 +1280,7 @@ final public class ProgramTree extends JPanel implements
 	    InfoPanel.searchButton.doClick();
 	    logger.info ("User rejected MSB");
 	}
-	if ( !anObservationHasBeenDone ) {
+	if ( !anObservationHasBeenDone && msbPendingFile != null ) {
 	    msbPendingFile.delete();
 	}
 	return done;
